backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
pymongo==4.7.2
python-dotenv==1.0.1
bcrypt==4.1.3
fpdf2
